Fix should_output_info_for_multiple_files
Test should_output_info_for_multiple_files fails because the output is
not in the same order as the input.  Add assert_equal_any_order() to
the OutputCollector, which sorts the expected and actual lines so that
the output matches even if the order in which they are shown in
trash-list is different.