'trash-put -f ...' now ignores files/dirs that do not exist.
Similar to 'rm -f ...', it is handy in scripts to ignore
files and directories that do not exist.
Often a script writer just wants to delete (or trash)
a file if it exists, without reporting any errors if
it does not exist.

It is difficult to correctly check for absent files
using wildcards from a bash shell.
See here for examples of the awkwardness of lacking -f :

https://stackoverflow.com/questions/6363441/check-if-a-file-exists-with-wildcard-in-shell-script

When -f is specified, any files or directories that do
not exist are skipped without causing any error messages
or nonzero status codes.

Updated help text printed by -h or --help to explain -f.

Updated to include a -f unit test.
Here is output from running all tests:

(env) don@spearmint:~/github/trash-cli$ nosetests
.............................................................................................................S..................................................................................................
----------------------------------------------------------------------
Ran 208 tests in 1.136s

OK (SKIP=1)